requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
deprecation==2.1.0
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Markdown==3.7
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.0
psycopg==3.2.3
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-decouple==3.8
realtime==2.0.6
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
tzdata==2024.2
websockets==13.1
yarl==1.18.0
